after rebase, move CircleProportionChart into nivo/NivoCharts.js
