--- client/src/charts/NivoCharts.yaml
@@ -1,10 +1,0 @@
-bubble_title:
-  transform: [handlebars, markdown]
-  en: |
-    **{{inner}}** as a fraction of **{{outer}}**
-  fr: |
-    TODO
-
-bubble_all_other:
-  en: All other
-  fr: TODO
